+++ client-v2/src/App.tsx
@@ -1,76 +1,12 @@
-import { Button } from "@/components/ui/button";
-
-import { Input } from "@/components/ui/input";
-import { MouseEventHandler, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 
 import { DailyChart } from "./DailyChart";
 import { StatsData } from "./data-types";
-import { ProjectChart } from "./ProjectChart";
 import { LanguageChart } from "./LanguageChart";
-
-const BASE_URL = "http://localhost:8080";
-
-function Login({ onLogin }: { onLogin: (t: string) => void }) {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [isLoading, setLoading] = useState(false);
-  const [errMsg, setErrMsg] = useState<string | null>(null);
-  const handleSubmit: MouseEventHandler<HTMLButtonElement> = async (e) => {
-    e.preventDefault();
-    setLoading(true);
-    setErrMsg(null);
-
-    try {
-      const res = await fetch(`${BASE_URL}/api/v1/login`, {
-        method: "POST",
-        body: JSON.stringify({
-          username,
-          password,
-        }),
-      });
-      const body = await res.json();
-      if (res.ok) {
-        onLogin(body?.token);
-        return;
-      }
-      throw new Error("failed to login");
-    } catch (err) {
-      console.log(err);
-      setErrMsg("Something went wrong");
-    } finally {
-      setLoading(false);
-    }
-  };
-  return (
-    <div className="w-full max-w-sm items-center">
-      <Input
-        value={username}
-        onChange={(e) => setUsername(e.target.value)}
-        className="mt-4"
-        type="username"
-        placeholder="Username"
-      />
-      <Input
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-        className="mt-4"
-        type="password"
-        placeholder="Password"
-      />
-      {errMsg && <p className="text-red-500">{errMsg}</p>}
-      {isLoading && <p className="text-slate-400 animate-pulse">Loading...</p>}
-      <Button
-        className="mt-4 w-1/4 float-end"
-        type="submit"
-        onClick={handleSubmit}
-      >
-        Login
-      </Button>
-    </div>
-  );
-}
-
-const TOKEN_KEY = "WakaToken";
+import { Login } from "./Login";
+import { ProjectChart } from "./ProjectChart";
+import { BASE_URL, DEFAULT_DAY_RANGE, getDateRange, TOKEN_KEY } from "./utils";
+import { Banner } from "./Banner";
 
 function App() {
   const [token, setToken] = useState<string | null>(null);
@@ -107,22 +43,6 @@ function App() {
       <Plot onLogout={onLogout} token={token} />
     </>
   );
-}
-
-const DEFAULT_DAY_RANGE = 7;
-
-function getDateRange(days: number) {
-  const formatDateForReq = (date: Date) => {
-    return date.toISOString().split("T")[0];
-  };
-
-  const endDate = new Date();
-  endDate.setDate(endDate.getDate() - 1);
-  const ending = formatDateForReq(endDate);
-  const startDate = new Date();
-  startDate.setDate(startDate.getDate() - days);
-  const starting = formatDateForReq(startDate);
-  return { starting, ending };
 }
 
 function Plot({ onLogout, token }: { onLogout: () => void; token: string }) {
@@ -174,17 +94,6 @@ function Plot({ onLogout, token }: { onLogout: () => void; token: string }) {
       {data?.projectStats && (
         <ProjectChart projectDurations={data?.projectStats?.durations} />
       )}
-    </div>
-  );
-}
-
-function Banner({ onLogout }: { onLogout: () => void }) {
-  return (
-    <div className="flex bg-slate-400 p-6 gap-8 items-center">
-      <h2 className="text-xl ">Waka Dashboard</h2>
-      <Button className="mr-4" onClick={onLogout}>
-        Logout
-      </Button>
     </div>
   );
 }
